requirements.txt
Flask
Werkzeug
flask_table
Flask-Bootstrap
flask_wtf
wtforms_components
